requirements.txt
wheel==0.38.4
lxml==4.9.2
PyQt5==5.15.7
PyQtWebEngine==5.15.6
Orange3==3.34.1
Orange3-Text==1.12.0
Orange3-network==1.7.0
pandas==1.5.3
regex==2022.10.31
stanza==1.4.2
spacy==3.4.4
dgl==0.6.1
sphinx==5.3.0
recommonmark==0.7.1
sphinx-rtd-theme==1.1.1
Cython==0.29.33
dhtmlparser3==3.0.17
-e git+https://github.com/kodymoodley/stroll-singlethreaded@master#egg=stroll-srl
